pyproject.toml
[tool.poetry]
name = "koutou"
version = "0.1.0"
description = "koutou scraper"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8"
selenium = "^3.141.0"
psycopg2 = "^2.8.5"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"